my_flask_app/requirements.txt
Flask
requests
easyocr
google-generativeai
google-cloud-translate
opencv-python
numpy
Pillow
python-dotenv
pytest
pytest-flask
